jwt add for units

+++ task-backend/src/main/java/lk/ijse/gdse/taskbackend/controllers/UnitController.java
@@ -1,5 +1,6 @@
 package lk.ijse.gdse.taskbackend.controllers;
 
+import lk.ijse.gdse.taskbackend.auth.JwtTokenProvider;
 import lk.ijse.gdse.taskbackend.dto.UnitDTO;
 import lk.ijse.gdse.taskbackend.entity.Unit;
 import lk.ijse.gdse.taskbackend.service.UnitService;
@@ -7,6 +8,7 @@ import lk.ijse.gdse.taskbackend.util.ResponseUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -14,49 +16,69 @@ import java.util.Optional;
 @CrossOrigin("*")
 @RequestMapping("/api/unit")
 public class UnitController {
-    @Autowired
-    private final UnitService unitService;
 
-    public UnitController(UnitService unitService) {
+    private final UnitService unitService;
+    private final JwtTokenProvider jwtTokenProvider;
+
+    @Autowired
+    public UnitController(UnitService unitService, JwtTokenProvider jwtTokenProvider) {
         this.unitService = unitService;
+        this.jwtTokenProvider = jwtTokenProvider;
     }
 
     @PostMapping
-    public ResponseUtil saveUnit(@RequestBody UnitDTO unitDTO) {
-        Unit unit = unitService.saveUnit(unitDTO);
-        return new ResponseUtil(200, "successfully save unit", unit);
+    public ResponseUtil saveUnit(@RequestBody UnitDTO unitDTO, @RequestHeader("Authorization") String authorizationHeader) {
+        if (jwtTokenProvider.validateToken(authorizationHeader)) {
+            Unit unit = unitService.saveUnit(unitDTO);
+            return new ResponseUtil(200, "successfully save unit", unit);
+        }
+        return new ResponseUtil(401, "UnAuthorization", null);
     }
 
     @PutMapping
-    public ResponseUtil updateUnit(@RequestBody UnitDTO unitDTO) {
-        Unit unit = unitService.updateUnit(unitDTO);
-        return new ResponseUtil(200, "successfully update unit", unit);
+    public ResponseUtil updateUnit(@RequestBody UnitDTO unitDTO, @RequestHeader("Authorization") String authorizationHeader) {
+        if (jwtTokenProvider.validateToken(authorizationHeader)) {
+            Unit unit = unitService.updateUnit(unitDTO);
+            return new ResponseUtil(200, "successfully update unit", unit);
+        }
+        return new ResponseUtil(401, "UnAuthorization", null);
     }
 
     @GetMapping
-    public List<Unit> getAllUnit() {
-        List<Unit> allUnit = unitService.getAllUnit();
-        //  return new ResponseUtil(200, "fount all unit", allUnit);
-        return allUnit;
+    public List<Unit> getAllUnit(@RequestHeader("Authorization") String authorizationHeader) {
+        if (jwtTokenProvider.validateToken(authorizationHeader)) {
+            return unitService.getAllUnit();
+            //  return new ResponseUtil(200, "fount all unit", allUnit);
+        }
+        return new ArrayList<>();
     }
 
     @GetMapping(path = "/getUnitById/{id}")
-    public Unit getUnitById(@PathVariable Long id) {
-        Optional<Unit> unitById = unitService.findUnitById(id);
+    public Unit getUnitById(@PathVariable Long id, @RequestHeader("Authorization") String authorizationHeader) {
+        if (jwtTokenProvider.validateToken(authorizationHeader)) {
+            Optional<Unit> unitById = unitService.findUnitById(id);
 //        return new ResponseUtil(200, "found unit", unitById);
-    return unitById.get();
+            return unitById.get();
+        }
+        return new Unit();
     }
 
 
     @DeleteMapping("/{id}")
-    public ResponseUtil deleteUnitById(@PathVariable Long id) {
-        unitService.deleteUnit(id);
-        return new ResponseUtil(200, "successfully Delete unit", null);
+    public ResponseUtil deleteUnitById(@PathVariable Long id, @RequestHeader("Authorization") String authorizationHeader) {
+        if (jwtTokenProvider.validateToken(authorizationHeader)) {
+            unitService.deleteUnit(id);
+            return new ResponseUtil(200, "successfully Delete unit", null);
+        }
+        return new ResponseUtil(401, "UnAuthorization", null);
     }
 
     @GetMapping(path = "/getIds")
-    public List<String> getAllNames() {
-        return unitService.findAllUnitIds();
+    public List<String> getAllNames(@RequestHeader("Authorization") String authorizationHeader) {
+        if (jwtTokenProvider.validateToken(authorizationHeader)) {
+            return unitService.findAllUnitIds();
+        }
+        return new ArrayList<>();
     }
 
 }
